AureusPublic/requirements.txt
aiohttp==3.7.3
altgraph==0.17
astunparse==1.6.3
async-timeout==3.0.1
asyncio-dgram==1.2.0
attrs==20.3.0
braceexpand==0.1.6
cffi==1.14.5
chardet==3.0.4
click==7.1.2
discord.py==1.6.0
dnspython==1.15.0
dnspython3==1.15.0
future==0.18.2
humanize==3.2.0
idna==3.1
import-expression==1.1.4
jishaku==1.20.0.220
mcstatus==5.1.1
multidict==5.1.0
pefile==2019.4.18
Pillow==8.1.0
pycparser==2.20
pyjokes==0.6.0
PyNaCl==1.4.0
python-dotenv==0.15.0
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
